+++ app/src/main/java/com/nextcloud/talk/attachmentpreview/FileDescription.kt
@@ -1,0 +1,29 @@
+/*
+ * Nextcloud Talk - Android Client
+ *
+ * SPDX-FileCopyrightText: 2026 Nextcloud GmbH and Nextcloud contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+package com.nextcloud.talk.attachmentpreview
+
+import android.graphics.Bitmap
+import com.nextcloud.talk.utils.Mimetype
+
+internal enum class MediaKind { IMAGE, VIDEO, OTHER }
+
+internal data class FileDescription(
+    val uri: String,
+    val name: String,
+    val kind: MediaKind,
+    val mimeType: String?,
+    val detail: String?,
+    val aspectRatio: Float? = null,
+    val videoThumbnail: Bitmap? = null
+)
+
+internal fun mediaKind(mimeType: String?): MediaKind =
+    when {
+        mimeType?.startsWith(Mimetype.IMAGE_PREFIX) == true -> MediaKind.IMAGE
+        mimeType?.startsWith(Mimetype.VIDEO_PREFIX) == true -> MediaKind.VIDEO
+        else -> MediaKind.OTHER
+    }
